cortana-deepseek-ocr/pyproject.toml
[project]
name = "cortana-deepseek-ocr"
version = "0.1.0"
description = "DeepSeek OCR inference microservice"
requires-python = ">=3.12"

dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "transformers==4.46.3",
  "torch==2.6.0",
  "huggingface-hub",
  "tokenizers==0.20.3",
  "einops",
  "addict",
  "easydict",
  "safetensors",
  "pillow",
  "pdf2image"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
